+++ b4n-tui/table/tabular_list.rs
@@ -172,9 +172,9 @@ impl<T: Row + Filterable<Fc>, Fc: FilterContext> TabularList<T, Fc> {
     pub fn get_mouse_menu_position(&self, line_no: u16, resource_name: &str, area: Rect) -> Position {
         let view = self.header.get_cached_view();
         let width = self.header.get_cached_width().unwrap_or_default();
-        let (namespace_width, name_width, extra_space) = self.header.get_widths(view, width);
-        let name_width = (name_width + extra_space).min(resource_name.chars().count());
-        let x = u16::try_from(namespace_width + name_width + 6).unwrap_or_default();
+        let widths = self.header.get_widths(view, width);
+        let name_width = (widths.name + widths.name_extra).min(resource_name.chars().count());
+        let x = u16::try_from(widths.group + name_width + 6).unwrap_or_default();
         let y = line_no.saturating_add(area.y);
 
         Position::new(x, y)