+++ b4n/ui/views/describe/mod.rs
@@ -1,0 +1,4 @@
+pub use self::view::*;
+
+mod content;
+mod view;