+++ b4n/ui/views/forwards/list.rs
@@ -78,6 +78,7 @@ impl Table for PortForwardsList {
             fn select_highlighted_item(&mut self);
             fn get_selected_items(&self) -> HashMap<&str, Vec<&str>>;
             fn is_anything_selected(&self) -> bool;
+            fn set_page(&mut self, page_start: usize, page_height: u16);
             fn update_page(&mut self, new_height: u16);
             fn get_paged_names(&self, width: usize) -> Vec<(String, bool)>;
         }
@@ -101,19 +102,12 @@ impl Table for PortForwardsList {
 
     /// Returns items from the current page in a form of text lines to display and colors for that lines.
     fn get_paged_items(&self, theme: &Theme, view: ViewType, width: usize) -> Vec<(String, TextColors)> {
-        let (namespace_width, name_width, name_extra_width) = self.table.header.get_widths(view, width);
+        let widths = self.table.header.get_widths(view, width);
 
         let mut result = Vec::with_capacity(self.table.list.page_height().into());
         for item in self.table.list.get_page() {
             result.push((
-                item.get_text(
-                    view,
-                    &self.table.header,
-                    width,
-                    namespace_width,
-                    name_width + name_extra_width,
-                    self.table.offset(),
-                ),
+                item.get_text(view, &self.table.header, &widths, width, self.table.offset()),
                 item.data.get_colors(theme, item.is_active, item.is_selected),
             ));
         }